branch: pos_v2 - new branch for Seperation of concerns project
